+++ src/QuickMarcEditor/constants.js
@@ -1,1 +1,8 @@
 export const LEADER_TAG = 'LDR';
+
+export const QUICK_MARC_ACTIONS = {
+  EDIT: 'edit',
+  DUPLICATE: 'duplicate',
+};
+
+export const FIELD_TAGS_TO_REMOVE = ['001', '005', '999'];